+++ Applications/PixelPaint/CMakeLists.txt
@@ -3,12 +3,6 @@ set(SOURCES
     CreateNewLayerDialog.cpp
     EllipseTool.cpp
     EraseTool.cpp
-    Filters/BoxBlurFilter.cpp
-    Filters/Filter.cpp
-    Filters/GenericConvolutionFilter.cpp
-    Filters/LaplacianFilter.cpp
-    Filters/SharpenFilter.cpp
-    Filters/SpatialGaussianBlurFilter.cpp
     Image.cpp
     ImageEditor.cpp
     Layer.cpp
