--- Applications/PixelPaint/Filters/Filter.h
@@ -1,66 +1,0 @@
-/*
- * Copyright (c) 2020, the SerenityOS developers.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#pragma once
-
-#include <LibGUI/Event.h>
-#include <LibGUI/Forward.h>
-#include <LibGfx/Forward.h>
-
-namespace PixelPaint {
-
-class Filter {
-public:
-    class Parameters {
-    public:
-        Parameters(Gfx::Bitmap& bitmap, const Gfx::IntRect& rect)
-            : m_target_bitmap(bitmap)
-            , m_target_rect(rect)
-
-        {
-        }
-
-        Gfx::Bitmap& bitmap() const { return m_target_bitmap; }
-        const Gfx::IntRect& rect() const { return m_target_rect; }
-        virtual bool is_generic_convolution_filter() const { return false; }
-
-        virtual ~Parameters() { }
-
-    private:
-        Gfx::Bitmap& m_target_bitmap;
-        Gfx::IntRect m_target_rect;
-    };
-    virtual ~Filter();
-
-    virtual const char* class_name() const = 0;
-
-    virtual void apply(const Parameters&) = 0;
-
-protected:
-    Filter();
-};
-
-}
