+++ Libraries/LibGfx/Filters/GenericConvolutionFilter.h
@@ -27,11 +27,10 @@
 #pragma once
 
 #include "Filter.h"
-#include <LibGUI/Dialog.h>
 #include <LibGfx/Matrix.h>
 #include <LibGfx/Matrix4x4.h>
 
-namespace PixelPaint {
+namespace Gfx {
 
 template<size_t N, typename T>
 inline static constexpr void normalize(Matrix<N, T>& matrix)
@@ -72,12 +71,66 @@ public:
         bool m_should_wrap { false };
     };
 
-    GenericConvolutionFilter();
-    virtual ~GenericConvolutionFilter();
+    GenericConvolutionFilter() { }
+    virtual ~GenericConvolutionFilter() { }
 
     virtual const char* class_name() const override { return "GenericConvolutionFilter"; }
 
-    virtual void apply(const Filter::Parameters&) override;
+    virtual void apply(const Filter::Parameters& parameters)
+    {
+        ASSERT(parameters.is_generic_convolution_filter());
+
+        auto& gcf_params = static_cast<const GenericConvolutionFilter::Parameters&>(parameters);
+
+        auto& source = gcf_params.bitmap();
+        const auto& source_rect = gcf_params.rect();
+        auto target = Gfx::Bitmap::create(source.format(), parameters.rect().size());
+
+        // FIXME: Help! I am naive!
+        for (auto i_ = 0; i_ < source_rect.width(); ++i_) {
+            auto i = i_ + source_rect.x();
+            for (auto j_ = 0; j_ < source_rect.height(); ++j_) {
+                auto j = j_ + source_rect.y();
+                FloatVector3 value(0, 0, 0);
+                for (auto k = 0; k < 4; ++k) {
+                    auto ki = i + k - 2;
+                    if (ki < 0 || ki >= source.size().width()) {
+                        if (gcf_params.should_wrap())
+                            ki = (ki + source.size().width()) % source.size().width();
+                        else
+                            continue;
+                    }
+
+                    for (auto l = 0; l < 4; ++l) {
+                        auto lj = j + l - 2;
+                        if (lj < 0 || lj >= source.size().height()) {
+                            if (gcf_params.should_wrap())
+                                lj = (lj + source.size().height()) % source.size().height();
+                            else
+                                continue;
+                        }
+
+                        auto pixel = source.get_pixel(ki, lj);
+                        FloatVector3 pixel_value(pixel.red(), pixel.green(), pixel.blue());
+
+                        value = value + pixel_value * gcf_params.kernel().elements()[k][l];
+                    }
+                }
+
+                // The float->u8 overflow is intentional.
+                target->set_pixel(i_, j_, Color(value.x(), value.y(), value.z(), source.get_pixel(i, j).alpha()));
+            }
+        }
+
+        // FIXME: Substitute for some sort of faster "blit" method.
+        for (auto i_ = 0; i_ < source_rect.width(); ++i_) {
+            auto i = i_ + source_rect.x();
+            for (auto j_ = 0; j_ < source_rect.height(); ++j_) {
+                auto j = j_ + source_rect.y();
+                source.set_pixel(i, j, target->get_pixel(i_, j_));
+            }
+        }
+    }
 };
 
 }
